pyproject.toml
[tool.poetry]
name = "booking"
version = "0.1.0"
description = ""
authors = ["AlexanderRemizov <alvrem@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.2"}
gunicorn = "^21.2.0"
sqlalchemy = "^2.0.21"
python-json-logger = "^2.0.7"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
asyncpg = "^0.28.0"
alembic = "^1.12.0"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
flower = "^2.0.1"
pillow = "^10.1.0"
celery = "5.2.7"
sqladmin = "^0.15.2"
bcrypt = "^4.0.1"
sentry-sdk = {extras = ["fastapi"], version = "^1.32.0"}
prometheus-fastapi-instrumentator = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
